pyproject.toml
[project]
name = "todo_django"
version = "0.1.0"
description = "A todo app based on https://dev.to/rodbv/creating-a-to-do-app-with-django-and-htmx"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
]

[dependency-groups]
dev = [
    "model-bakery>=1.20.1",
    "pytest-django>=4.9.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "todo_django.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

# Add verbose output
addopts = "-v"
